factor out DVS PACT export
