+++ backends/dory/__init__.py
@@ -1,2 +1,3 @@
-from .pact_export import export_net, export_dvsnet
+from .pact_export import export_net
+from .dvs_pact_export import export_dvsnet
 from .dory_passes import *
